requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch==2.7.1
openai-whisper
